make use of the UTimer Control.
add a property indicating whether the timer is paused or running.
