+++ SnakeEater/Common/Consts.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SnakeEater.Common
+{
+    /// <summary>
+    /// Constants class
+    /// </summary>
+    public class Consts
+    {
+        #region Consts
+        /// <summary>
+        /// The max x position of snake or food.
+        /// </summary>
+        public const int MAX_X = 64;
+
+        /// <summary>
+        /// The max y position of snake or food.
+        /// </summary>
+        public const int MAX_Y = 40;
+        #endregion
+    }
+}
